requirements_mobile.txt
# 移动端专用依赖包
# 核心框架
kivy==2.1.0
kivymd==1.1.1

# 网络请求
requests==2.31.0
urllib3==1.26.18

# Telegram API (轻量版)
telethon==1.34.0
cryptg==0.4.0

# 配置管理
python-dotenv==1.0.0
pydantic==2.5.2

# 异步处理
asyncio
aiofiles==23.2.1

# 数据处理
python-dateutil==2.8.2

# Android特定
pyjnius==1.4.2
plyer==2.1.0

# 加密和安全
cryptography==41.0.8
pyopenssl==23.3.0

# JSON处理
ujson==5.8.0
